Refactor main
