+++ examples/05-bad-set-example.yaml
@@ -2,7 +2,8 @@ rules:
   - cre:
       id: bad-order
     metadata:
-      id: rule-id
+      id: UoddCfP6YFxLPpX9BdmYdh
+      hash: BhzLAwg1SxZpX3mYUacw2g
     rule:
       set:
         event:
