+++ examples/20-bad-regex-example.yaml
@@ -2,7 +2,8 @@ rules:
   - cre:
       id: bad-regex-example
     metadata:
-      id: rule-id
+      id: A4ZFQZLprRMYH6g9kfiJRe
+      hash: B5MsHJocMTimFUzEU1isyP
     rule:
       set:
         event:
